benchmark: Make all sign conversions explicit

+++ benchmark/libponyc/common/common.cc
@@ -5,7 +5,7 @@ static void $$$$$$_PauseResumeOverheadOnce(benchmark::State& st) {
     st.PauseTiming();
     st.ResumeTiming();
   }
-  st.SetItemsProcessed(st.iterations());
+  st.SetItemsProcessed((int64_t)st.iterations());
 }
 
 BENCHMARK($$$$$$_PauseResumeOverheadOnce);
@@ -17,7 +17,7 @@ static void $$$$$$_PauseResumeOverheadTwice(benchmark::State& st) {
     st.PauseTiming();
     st.ResumeTiming();
   }
-  st.SetItemsProcessed(st.iterations());
+  st.SetItemsProcessed((int64_t)st.iterations());
 }
 
 BENCHMARK($$$$$$_PauseResumeOverheadTwice);
@@ -31,7 +31,7 @@ static void $$$$$$_PauseResumeOverheadThrice(benchmark::State& st) {
     st.PauseTiming();
     st.ResumeTiming();
   }
-  st.SetItemsProcessed(st.iterations());
+  st.SetItemsProcessed((int64_t)st.iterations());
 }
 
 BENCHMARK($$$$$$_PauseResumeOverheadThrice);
